% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectLymphocyte.R
\docType{class}
\name{TrenaProjectLymphocyte-class}
\alias{TrenaProjectLymphocyte-class}
\alias{.TrenaProjectLymphocyte}
\alias{TrenaProjectLymphocyte}
\title{TrenaProjectLymphocyte-class}
\usage{
TrenaProjectLymphocyte(quiet = TRUE)
}
\value{
An object of the TrenaProjectLymphocyte class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
